requirements.txt
# requirements.txt
# 核心依赖
# deepseekapi==0.1.0      # 您已实现的API客户端
openai>=1.3.0
requests>=2.26.0
pydantic>=1.9.0
python-dotenv>=0.19.0
PyQt5>=5.15             # 跨平台GUI框架
pynput>=1.7             # 全局热键监听
mss>=9.0                # 高性能截屏
pytesseract>=0.3.10     # OCR文字识别
pyclip>=0.7.0           # 剪贴板操作
numpy>=1.24             # 图像处理

# Windows专用
pywin32>=306; sys_platform == 'win32'

# Linux专用
python-xlib>=0.33; sys_platform == 'linux'